fix conflict problem & move socket.io in app.js to config socketIo.js
